chore: separate automated tests and manual HTTP files into test/ and http/
