+++ server/http/auth/logout.http
