+++ server/test/helpers.ts
@@ -1,6 +1,6 @@
 import request from 'supertest';
 
-import { app } from '../app';
+import { app } from '../src/app';
 
 export const validUser = {
   username: 'alice',
