+++ server/tsconfig.json
@@ -27,6 +27,6 @@
       "@/*": ["./src/*"]
     },
   },
-  "include": ["src"],
+  "include": ["src", "test"],
   "exclude": ["dist", "node_modules"],
 }
